Rename all book elements to product
